If the authority server goes down, the oidc.security.check-session service will now emit an onCheckSessionChanged event.
The server will be checked every sixty seconds.
If a message isn't received from the check session iframe within 3 "beats" of the check session heart beat. An event will be emitted.

The iframe will only stop emitting events after the server is checked. So the event will be emitted at a minimum of 3 beats (9 seconds) and  maximum of sixty seconds + 3 beats (69 seconds).

Fixes #310