Optional url handler for logoff function
Authorization function has the option to pass a function that hanlde the url instead of redirect to it. I propose that logoff function works the same way and allow the user to pass a function that handle the url as it want or redirect if there is not handler in parameters.