Use window object safely injecting DOCUMENT
Using the global object window in teh server side produces errors, and it is discouraged.

The correct approach is to use the defaultView of the DOCUMENT injecting it.

Some features such as storage and crypto may not be available in the server, so we need to check before using them.

It is fine togo this way since there is no storage in the backend, thus there is no authentication happening there.